requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
playwright>=1.44.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
